Fix up some test locations. Attend to a copilot comment
