+++ lib/handlers/noscript.ts
@@ -33,12 +33,9 @@ import {logger} from '../logger.js';
 import {ClientOptionsHandler} from '../options-handler.js';
 import {StorageBase} from '../storage/index.js';
 
+import {isMobileViewer} from '../app/url-handlers.js';
 import {RenderConfig} from './handler.interfaces.js';
 import {cached, csp} from './middleware.js';
-
-function isMobileViewer(req: express.Request) {
-    return req.header('CloudFront-Is-Mobile-Viewer') === 'true';
-}
 
 export class NoScriptHandler {
     constructor(
